+++ MyHelper.Api/DAL/Entities/Note.cs
@@ -18,7 +18,7 @@ namespace MyHelper.Api.DAL.Entities
 
         public DateTime UpdateDate { get; set; }
 
-        public int AppUserId { get; set; }
+        public long AppUserId { get; set; }
 
         public AppUser AppUser { get; set; }
